+++ src/com/ttt/ai/hal/Hal.java
@@ -1,4 +1,4 @@
-package com.ttt.ai;
+package com.ttt.ai.hal;
 
 import java.util.HashMap;
